+++ CHANGELOG.md
@@ -1,5 +1,9 @@
-### Create: Classic Blaze Enchanter 1.0.3
+### Create: Classic Blaze Enchanter 1.1.0
 
 #### Update
-* Ponder has been improved.
-* Add Japanese Translation by @YukkuriOfuton (due to ponder change, it's incomplete)
+* Classic Blaze Enchanter supports direct belt input and output
+* Add Ponder for automation
+* Add Russian Translation by @Korben
+
+#### Fix
+* Fix Blaze's Enchanting Handbook not show up in Creative Tab